Trying again to fix home tap
The page URL structure is different on GitHub and complicates the
resolution of URLs according to the jekyll manual[0]. Use the
`relative_url` filter with the `baseurl`.

Why not `url`? Well if we do that then local urls break and we end up
with something like `http://127.0.0.1:4000//http://127.0.0.1:4000/` when
tapping the home logo or heading.

[0]: https://jekyllrb.com/docs/github-pages/