Change default site.url to use HTTPS URL of the CNAME
Another attempt at fixing the URLs for tags and tapping home button.

I thought this was fixed in

* 1206d0764a1d (Fix link for tags on tech.finn.no/tags, 2018-11-24)
* a5d6d9fbd169 (Trying again to fix home tap, 2018-11-24)

but when visiting the blog it did not work.

I am still not sure if it will work but this time I tested this on one
of my [own domains][0] using a different CNAME. It worked there and
localhost is still functional.

[0]: tech.finn.xdp.no